+++ install_from_dist.sh
@@ -2,19 +2,19 @@
 
 set -e
 
-sudo mkdir -p /usr/local/bin
-sudo mkdir -p /usr/local/share/radiobrowser
+sudo mkdir -p /usr/bin
+sudo mkdir -p /usr/share/radiobrowser
 sudo mkdir -p /var/log/radiobrowser
 
-sudo cp target/release/radiobrowser-api-rust /usr/local/bin/radiobrowser
+sudo cp target/release/radiobrowser-api-rust /usr/bin/radiobrowser
 sudo cp init/radiobrowser.service /etc/systemd/system
-sudo cp static/* /usr/local/share/radiobrowser/
+sudo cp static/* /usr/share/radiobrowser/
 sudo cp etc/config-example.toml /etc/radiobrowser/config-example.toml
 if [ ! -f /etc/radiobrowser/config.toml ]; then
     sudo cp etc/config-example.toml /etc/radiobrowser/config.toml
 fi
 
-sudo chmod ugo+x /usr/local/bin/radiobrowser
+sudo chmod ugo+x /usr/bin/radiobrowser
 sudo groupadd --system radiobrowser
 sudo useradd --system --no-create-home --home-dir /var/lib/radiobrowser --gid radiobrowser radiobrowser