+++ src/db/db.rs
@@ -10,6 +10,7 @@ use crate::db::models::StationClickItem;
 use crate::db::MysqlConnection;
 use crate::db::DbError;
 use std::error::Error;
+use std::collections::HashSet;
 use std::collections::HashMap;
 
 pub trait DbConnection {
@@ -61,7 +62,7 @@ pub trait DbConnection {
     fn get_checks(&self, stationuuid: Option<String>, checkuuid: Option<String>, seconds: u32, include_history: bool) -> Result<Vec<StationCheckItem>, Box<dyn Error>>;
     fn get_clicks(&self, stationuuid: Option<String>, clickuuid: Option<String>, seconds: u32) -> Result<Vec<StationClickItem>, Box<dyn Error>>;
 
-    fn insert_checks(&self, list: &Vec<StationCheckItemNew>) -> Result<(), Box<dyn Error>>;
+    fn insert_checks(&self, list: &Vec<StationCheckItemNew>) -> Result<HashSet<String>, Box<dyn std::error::Error>>;
     fn update_station_with_check_data(&self, list: &Vec<StationCheckItemNew>, local: bool) -> Result<(), Box<dyn Error>>;
 
     fn insert_clicks(&self, list: &Vec<StationClickItemNew>) -> Result<(), Box<dyn Error>>;
@@ -70,6 +71,7 @@ pub trait DbConnection {
     fn delete_were_working(&mut self, hours: u32) -> Result<(), Box<dyn Error>>;
     fn delete_old_checks(&mut self, hours: u32) -> Result<(), Box<dyn Error>>;
     fn delete_old_clicks(&mut self, hours: u32) -> Result<(), Box<dyn Error>>;
+    fn remove_unused_ip_infos_from_stationclicks(&mut self, hours: u32) -> Result<(), Box<dyn Error>>;
 
     fn update_stations_clickcount(&self) -> Result<(), Box<dyn Error>>;
 
@@ -80,7 +82,7 @@ pub trait DbConnection {
     fn remove_from_cache(&self, tags: Vec<&String>, table_name: &str, column_name: &str) -> Result<(), Box<dyn Error>>;
 
     fn vote_for_station(&self, ip: &str, station: Option<StationItem>) -> Result<String, Box<dyn Error>>;
-    fn increase_clicks(&self, ip: &str, station: &StationItem) -> Result<bool,Box<dyn Error>>;
+    fn increase_clicks(&self, ip: &str, station: &StationItem, hours: u32) -> Result<bool,Box<dyn Error>>;
 }
 
 pub fn connect(connection_string: String) -> Result<Box<dyn DbConnection>, Box<dyn std::error::Error>> {